test/smoke: extract runner to own file
